Fix: blog errors on deploy
