+++ lib/blog/index.ts
@@ -1,6 +1,11 @@
 /**
  * Blog Module Exports
+ * @module lib/blog
+ * @description
+ * Central export point for all blog functionality.
+ * Consolidates exports from core, MDX handling, and validation.
  */
 
+export * from './core';
 export * from './mdx';
 export * from './validation';
